🥅 [app.system.doc] => robustify `get_classes_inheriting_from`
The error appeared only in nightly build when youwol is installed using pipx, tarball or wheel. Somehow, in these cases, `DocCache.flat_classes` include types that can not be used in `issubclass`.

Not sure why this behavior does not appear when youwol is run from sources directly, the `try/except` is still relevant as a way to robustify the code.