♻️ [app/CDN] Rework CDN packaging for enhanced flexibility and efficiency
When creating a package for the CDN, we've introduced additional flexibility for the client. This update allows for selective file encoding (using Brotli) and the precise definition of content types for individual files.

Key Changes:

* Selective Encoding: Files can now be encoded with Brotli compression based on client specification.
* Content Type Definition: Clients can define specific content types for each file in the package.

These changes ensure that compression is performed once at the time of package creation, reducing the need for repeated compression during subsequent downloads by other users.

TG-2358
TG-2353